requirements.txt
PyOpenGL 
PyOpenGL_accelerate
numpy
PyQt5
